+++ res/Hydra Game File/Room/Monster 2.txt
@@ -1,0 +1,11 @@
+Room Name:
+Hazardous Environments Training Course
+Room ID:
+01
+Room Coordinate:
+0,0,4
+Room Description:
+A breezy Helicopter landing pad with two metal doors. The landing pad sits on top of the building and the lights of govern-ment vehicles can be seen below.
+Room Connection:
+East-02
+South-03
